Separate preview error snackbar into its own component
